Implemented Notifications for the DSC Settings Gatt Characteristic.
After connection and service discovery is complete, we wait 3sec and then Gatt Read Settings and get json payload.

The Gatt Server will respond within 1sec with and push (notify) the Settings back to here to validate.

We check the value against the SharedPreference model to make sure they match.